requirements.txt
pygame==1.9.6
matplotlib==3.2.1
seaborn==0.10.1
pandas==1.0.3
pillow==7.1.2
